+++ polaris-plugins/polaris-plugins-configfilter/src/main/java/com/tencent/polaris/plugins/configfilter/util/AESUtil.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.tencent.polaris.plugins.configfilter.crypto.util;
+package com.tencent.polaris.plugins.configfilter.util;
 
 import com.tencent.polaris.api.exception.ErrorCode;
 import com.tencent.polaris.api.exception.PolarisException;
